クラスを分けました
